# nfs-dashboard-backend/swagger.yaml
openapi: 3.0.3
info:
  title: NFS Dashboard API
  version: 1.0.0
  description: API documentation for the NFS Dashboard backend
servers:
  - url: http://localhost:8080
paths:
...
  /api/files:
    get:
      summary: List files and folders
      parameters:
        - in: query
          name: path
          schema:
            type: string
          required: true
      responses:
        '200':
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/File'
    delete:
      summary: Delete a file or folder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
              required: [path]
      responses:
        '204':
          description: Deleted successfully
        '400':
          description: Bad request
...
components:
  schemas:
...
    File:
      type: object
      required: [name, path, is_dir]
      properties:
        name:
          type: string
        path:
          type: string
        is_dir:
          type: boolean
        size:
          type: integer
        lastModified:
          type: string
          format: date-time
...